+++ model/site_category.go
@@ -2,6 +2,6 @@ package model
 
 type SiteCategory struct {
 	Model
-	Name        string `json:"name"`
-	SyncNodeIds []int  `json:"sync_node_ids" gorm:"serializer:json"`
+	Name        string   `json:"name"`
+	SyncNodeIds []uint64 `json:"sync_node_ids" gorm:"serializer:json"`
 }
